refactor into a seperate toolset
